requirements.txt
streamlit>=1.28.0
openai>=1.0.0
networkx>=3.0
pyvis>=0.3.2
pandas>=2.0.0
pyyaml>=6.0
plotly>=5.15.0 